+++ src/wave/heroes/nicolas/mod.rs
@@ -1,0 +1,2 @@
+use super::test_hero;
+test_hero!(Nicholas);