+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Hero Lydia (w/o Jolly Water fight)
+
 ## [0.0.17] - 2023-10-28
 
 ### Added